requirements.txt
flask==3.0.2
flask-cors==4.0.0
huggingface-hub==0.21.3
python-dotenv==1.0.0
gunicorn==21.2.0 